refactoring, fix error
Розділив на компоненти,
Змінив назви карток
Змінив значення вхідних даних
Додав селектор
